+++ packages/google_mobile_ads/android/src/main/java/io/flutter/plugins/googlemobileads/FlutterAdManagerAdRequest.java
@@ -27,9 +27,9 @@ import java.util.Objects;
  */
 class FlutterAdManagerAdRequest extends FlutterAdRequest {
 
-  @Nullable private Map<String, String> customTargeting;
-  @Nullable private Map<String, List<String>> customTargetingLists;
-  @Nullable private String publisherProvidedId;
+  @Nullable private final Map<String, String> customTargeting;
+  @Nullable private final Map<String, List<String>> customTargetingLists;
+  @Nullable private final String publisherProvidedId;
 
   static class Builder extends FlutterAdRequest.Builder {
 
@@ -64,6 +64,8 @@ class FlutterAdManagerAdRequest extends FlutterAdRequest {
           getHttpTimeoutMillis(),
           getLocation(),
           publisherProvidedId,
+          getMediationExtrasIdentifier(),
+          getMediationNetworkExtrasProvider(),
           getAdMobExtras());
     }
   }
@@ -78,6 +80,8 @@ class FlutterAdManagerAdRequest extends FlutterAdRequest {
       @Nullable Integer httpTimeoutMillis,
       @Nullable Location location,
       @Nullable String publisherProvidedId,
+      @Nullable String mediationExtrasIdentifier,
+      @Nullable MediationNetworkExtrasProvider mediationNetworkExtrasProvider,
       @Nullable Map<String, String> adMobExtras) {
     super(
         keywords,
@@ -86,23 +90,18 @@ class FlutterAdManagerAdRequest extends FlutterAdRequest {
         neighboringContentUrls,
         httpTimeoutMillis,
         location,
+        mediationExtrasIdentifier,
+        mediationNetworkExtrasProvider,
         adMobExtras);
     this.customTargeting = customTargeting;
     this.customTargetingLists = customTargetingLists;
     this.publisherProvidedId = publisherProvidedId;
   }
 
-  AdManagerAdRequest asAdManagerAdRequest() {
+  AdManagerAdRequest asAdManagerAdRequest(String adUnitId) {
     final AdManagerAdRequest.Builder builder = new AdManagerAdRequest.Builder();
-    if (getKeywords() != null) {
-      for (final String keyword : getKeywords()) {
-        builder.addKeyword(keyword);
-      }
-    }
+    updateAdRequestBuilder(builder, adUnitId);
 
-    if (getContentUrl() != null) {
-      builder.setContentUrl(getContentUrl());
-    }
     if (customTargeting != null) {
       for (final Map.Entry<String, String> entry : customTargeting.entrySet()) {
         builder.addCustomTargeting(entry.getKey(), entry.getValue());
@@ -113,14 +112,9 @@ class FlutterAdManagerAdRequest extends FlutterAdRequest {
         builder.addCustomTargeting(entry.getKey(), entry.getValue());
       }
     }
-    addNetworkExtras(builder);
     if (publisherProvidedId != null) {
       builder.setPublisherProvidedId(publisherProvidedId);
     }
-    if (getLocation() != null) {
-      builder.setLocation(getLocation());
-    }
-    builder.setRequestAgent(Constants.REQUEST_AGENT_PREFIX_VERSIONED);
     return builder.build();
   }