+++ packages/google_mobile_ads/android/src/main/java/io/flutter/plugins/googlemobileads/FlutterAdRequest.java
@@ -19,8 +19,11 @@ import android.os.Bundle;
 import androidx.annotation.Nullable;
 import com.google.ads.mediation.admob.AdMobAdapter;
 import com.google.android.gms.ads.AdRequest;
+import com.google.android.gms.ads.mediation.MediationExtrasReceiver;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Map.Entry;
 import java.util.Objects;
 
 class FlutterAdRequest {
@@ -30,6 +33,8 @@ class FlutterAdRequest {
   @Nullable private final List<String> neighboringContentUrls;
   @Nullable private final Integer httpTimeoutMillis;
   @Nullable private final Location location;
+  @Nullable private final String mediationExtrasIdentifier;
+  @Nullable private final MediationNetworkExtrasProvider mediationNetworkExtrasProvider;
   @Nullable private final Map<String, String> adMobExtras;
 
   protected static class Builder {
@@ -39,6 +44,8 @@ class FlutterAdRequest {
     @Nullable private List<String> neighboringContentUrls;
     @Nullable private Integer httpTimeoutMillis;
     @Nullable private Location location;
+    @Nullable private String mediationExtrasIdentifier;
+    @Nullable private MediationNetworkExtrasProvider mediationNetworkExtrasProvider;
     @Nullable private Map<String, String> adMobExtras;
 
     Builder setKeywords(@Nullable List<String> keywords) {
@@ -68,6 +75,17 @@ class FlutterAdRequest {
 
     Builder setLocation(@Nullable Location location) {
       this.location = location;
+      return this;
+    }
+
+    Builder setMediationNetworkExtrasIdentifier(@Nullable String mediationExtrasIdentifier) {
+      this.mediationExtrasIdentifier = mediationExtrasIdentifier;
+      return this;
+    }
+
+    Builder setMediationNetworkExtrasProvider(
+        @Nullable MediationNetworkExtrasProvider mediationNetworkExtrasProvider) {
+      this.mediationNetworkExtrasProvider = mediationNetworkExtrasProvider;
       return this;
     }
 
@@ -107,6 +125,16 @@ class FlutterAdRequest {
     }
 
     @Nullable
+    protected String getMediationExtrasIdentifier() {
+      return mediationExtrasIdentifier;
+    }
+
+    @Nullable
+    protected MediationNetworkExtrasProvider getMediationNetworkExtrasProvider() {
+      return mediationNetworkExtrasProvider;
+    }
+
+    @Nullable
     protected Map<String, String> getAdMobExtras() {
       return adMobExtras;
     }
@@ -119,6 +147,8 @@ class FlutterAdRequest {
           neighboringContentUrls,
           httpTimeoutMillis,
           location,
+          mediationExtrasIdentifier,
+          mediationNetworkExtrasProvider,
           adMobExtras);
     }
   }
@@ -130,6 +160,8 @@ class FlutterAdRequest {
       @Nullable List<String> neighboringContentUrls,
       @Nullable Integer httpTimeoutMillis,
       @Nullable Location location,
+      @Nullable String mediationExtrasIdentifier,
+      @Nullable MediationNetworkExtrasProvider mediationNetworkExtrasProvider,
       @Nullable Map<String, String> adMobExtras) {
     this.keywords = keywords;
     this.contentUrl = contentUrl;
@@ -137,28 +169,44 @@ class FlutterAdRequest {
     this.neighboringContentUrls = neighboringContentUrls;
     this.httpTimeoutMillis = httpTimeoutMillis;
     this.location = location;
+    this.mediationExtrasIdentifier = mediationExtrasIdentifier;
+    this.mediationNetworkExtrasProvider = mediationNetworkExtrasProvider;
     this.adMobExtras = adMobExtras;
   }
 
   /** Adds network extras to the ad request builder, if any. */
-  protected void addNetworkExtras(AdRequest.Builder builder) {
-    Bundle bundle = new Bundle();
-    if (adMobExtras != null) {
+  private void addNetworkExtras(AdRequest.Builder builder, String adUnitId) {
+    Map<Class<? extends MediationExtrasReceiver>, Bundle> networkExtras = new HashMap<>();
+    if (mediationNetworkExtrasProvider != null) {
+      Map<Class<? extends MediationExtrasReceiver>, Bundle> providedExtras =
+          mediationNetworkExtrasProvider.getMediationExtras(adUnitId, mediationExtrasIdentifier);
+      networkExtras.putAll(providedExtras);
+    }
+
+    if (adMobExtras != null && !adMobExtras.isEmpty()) {
+      Bundle adMobBundle = new Bundle();
       for (Map.Entry<String, String> extra : adMobExtras.entrySet()) {
-        bundle.putString(extra.getKey(), extra.getValue());
+        adMobBundle.putString(extra.getKey(), extra.getValue());
       }
+      networkExtras.put(AdMobAdapter.class, adMobBundle);
     }
+
     if (nonPersonalizedAds != null && nonPersonalizedAds) {
-      bundle.putString("npa", "1");
+      Bundle adMobBundle = networkExtras.get(AdMobAdapter.class);
+      if (adMobBundle == null) {
+        adMobBundle = new Bundle();
+      }
+      adMobBundle.putString("npa", "1");
+      networkExtras.put(AdMobAdapter.class, adMobBundle);
     }
-    if (!bundle.isEmpty()) {
-      builder.addNetworkExtrasBundle(AdMobAdapter.class, bundle);
+
+    for (Entry<Class<? extends MediationExtrasReceiver>, Bundle> entry : networkExtras.entrySet()) {
+      builder.addNetworkExtrasBundle(entry.getKey(), entry.getValue());
     }
   }
 
-  AdRequest asAdRequest() {
-    final AdRequest.Builder builder = new AdRequest.Builder();
-
+  /** Updates the {@link AdRequest.Builder} with the properties in this {@link FlutterAdRequest}. */
+  protected AdRequest.Builder updateAdRequestBuilder(AdRequest.Builder builder, String adUnitId) {
     if (keywords != null) {
       for (final String keyword : keywords) {
         builder.addKeyword(keyword);
@@ -167,7 +215,7 @@ class FlutterAdRequest {
     if (contentUrl != null) {
       builder.setContentUrl(contentUrl);
     }
-    addNetworkExtras(builder);
+    addNetworkExtras(builder, adUnitId);
     if (neighboringContentUrls != null) {
       builder.setNeighboringContentUrls(neighboringContentUrls);
     }
@@ -178,7 +226,11 @@ class FlutterAdRequest {
       builder.setLocation(location);
     }
     builder.setRequestAgent(Constants.REQUEST_AGENT_PREFIX_VERSIONED);
-    return builder.build();
+    return builder;
+  }
+
+  AdRequest asAdRequest(String adUnitId) {
+    return updateAdRequestBuilder(new AdRequest.Builder(), adUnitId).build();
   }
 
   @Nullable
@@ -212,6 +264,11 @@ class FlutterAdRequest {
   }
 
   @Nullable
+  protected String getMediationExtrasIdentifier() {
+    return mediationExtrasIdentifier;
+  }
+
+  @Nullable
   protected Map<String, String> getAdMobExtras() {
     return adMobExtras;
   }
@@ -237,6 +294,8 @@ class FlutterAdRequest {
                 && location.getLongitude() == request.location.getLongitude()
                 && location.getLatitude() == request.location.getLatitude()
                 && location.getTime() == request.location.getTime()))
+        && Objects.equals(mediationExtrasIdentifier, request.mediationExtrasIdentifier)
+        && Objects.equals(mediationNetworkExtrasProvider, request.mediationNetworkExtrasProvider)
         && Objects.equals(adMobExtras, request.adMobExtras);
   }
 
@@ -248,6 +307,8 @@ class FlutterAdRequest {
         nonPersonalizedAds,
         neighboringContentUrls,
         httpTimeoutMillis,
-        location);
+        location,
+        mediationExtrasIdentifier,
+        mediationNetworkExtrasProvider);
   }
 }