+++ packages/google_mobile_ads/ios/Tests/FLTGoogleMobileAdsReaderWriterTest.m
@@ -229,6 +229,7 @@
   NSArray<NSString *> *contentURLs = @[ @"url-1.com", @"url-2.com" ];
   request.neighboringContentURLs = contentURLs;
   request.location = [[FLTLocationParams alloc] initWithAccuracy:@1.5 longitude:@52 latitude:@123];
+  request.mediationExtrasIdentifier = @"identifier";
   request.adMobExtras = @{@"key" : @"value"};
   NSData *encodedMessage = [_messageCodec encode:request];
 
@@ -240,6 +241,7 @@
   XCTAssertEqualObjects(decodedRequest.location.accuracy, @1.5);
   XCTAssertEqualObjects(decodedRequest.location.longitude, @52);
   XCTAssertEqualObjects(decodedRequest.location.latitude, @123);
+  XCTAssertEqualObjects(decodedRequest.mediationExtrasIdentifier, @"identifier");
   XCTAssertEqualObjects(decodedRequest.adMobExtras, @{@"key" : @"value"});
 }
 
@@ -254,6 +256,7 @@
   request.neighboringContentURLs = contentURLs;
   request.pubProvidedID = @"pub-id";
   request.location = [[FLTLocationParams alloc] initWithAccuracy:@1.5 longitude:@52 latitude:@123];
+  request.mediationExtrasIdentifier = @"identifier";
   request.adMobExtras = @{@"key" : @"value"};
   NSData *encodedMessage = [_messageCodec encode:request];
 
@@ -269,6 +272,7 @@
   XCTAssertEqualObjects(decodedRequest.location.accuracy, @1.5);
   XCTAssertEqualObjects(decodedRequest.location.longitude, @52);
   XCTAssertEqualObjects(decodedRequest.location.latitude, @123);
+  XCTAssertEqualObjects(decodedRequest.mediationExtrasIdentifier, @"identifier");
   XCTAssertEqualObjects(decodedRequest.adMobExtras, @{@"key" : @"value"});
 }