+++ packages/google_mobile_ads/lib/src/ad_instance_manager.dart
@@ -737,6 +737,7 @@ class AdMessageCodec extends StandardMessageCodec {
       }
       writeValue(buffer, value.publisherProvidedId);
       writeValue(buffer, value.location);
+      writeValue(buffer, value.mediationExtrasIdentifier);
       writeValue(buffer, value.extras);
     } else if (value is AdRequest) {
       buffer.putUint8(_valueAdRequest);
@@ -748,6 +749,7 @@ class AdMessageCodec extends StandardMessageCodec {
         writeValue(buffer, value.httpTimeoutMillis);
       }
       writeValue(buffer, value.location);
+      writeValue(buffer, value.mediationExtrasIdentifier);
       writeValue(buffer, value.extras);
     } else if (value is RewardItem) {
       buffer.putUint8(_valueRewardItem);
@@ -887,6 +889,7 @@ class AdMessageCodec extends StandardMessageCodec {
               ? readValueOfType(buffer.getUint8(), buffer)
               : null,
           location: readValueOfType(buffer.getUint8(), buffer),
+          mediationExtrasIdentifier: readValueOfType(buffer.getUint8(), buffer),
           extras: readValueOfType(buffer.getUint8(), buffer)
               ?.cast<String, String>(),
         );
@@ -938,6 +941,7 @@ class AdMessageCodec extends StandardMessageCodec {
               : null,
           publisherProvidedId: readValueOfType(buffer.getUint8(), buffer),
           location: readValueOfType(buffer.getUint8(), buffer),
+          mediationExtrasIdentifier: readValueOfType(buffer.getUint8(), buffer),
           extras: readValueOfType(buffer.getUint8(), buffer)
               ?.cast<String, String>(),
         );