+++ packages/google_mobile_ads/test/ad_containers_test.dart
@@ -1329,6 +1329,7 @@ void main() {
         httpTimeoutMillis: 12345,
         location:
             LocationParams(accuracy: 1.1, longitude: 25, latitude: 38, time: 1),
+        mediationExtrasIdentifier: 'identifier',
         extras: {'key': 'value'},
       );
 
@@ -1347,6 +1348,7 @@ void main() {
         httpTimeoutMillis: 12345,
         location:
             LocationParams(accuracy: 1.1, longitude: 25, latitude: 38, time: 1),
+        mediationExtrasIdentifier: 'identifier',
         extras: {'key': 'value'},
       );
 
@@ -1357,9 +1359,11 @@ void main() {
       expect(decoded.contentUrl, adRequest.contentUrl);
       expect(decoded.nonPersonalizedAds, adRequest.nonPersonalizedAds);
       expect(decoded.keywords, adRequest.keywords);
+      // Time is not included on iOS.
       expect(decoded.location!.accuracy, 1.1);
       expect(decoded.location!.longitude, 25);
       expect(decoded.location!.latitude, 38);
+      expect(decoded.mediationExtrasIdentifier, 'identifier');
     });
 
     test('encode/decode $LoadAdError', () async {
@@ -1496,7 +1500,7 @@ void main() {
       );
     });
 
-    test('encode/decode AdManagerAdRequest Android', () async {
+    test('encode/decode $AdManagerAdRequest Android', () async {
       debugDefaultTargetPlatformOverride = TargetPlatform.android;
       final AdManagerAdRequest request = AdManagerAdRequest(
         keywords: <String>['who'],
@@ -1511,6 +1515,7 @@ void main() {
         publisherProvidedId: 'test-pub-id',
         location:
             LocationParams(accuracy: 1.1, longitude: 25, latitude: 38, time: 1),
+        mediationExtrasIdentifier: 'identifier',
         extras: {'key': 'value'},
       );
       final ByteData byteData = codec.encodeMessage(request)!;
@@ -1521,7 +1526,7 @@ void main() {
       );
     });
 
-    test('encode/decode AdManagerAdRequest iOS', () async {
+    test('encode/decode $AdManagerAdRequest iOS', () async {
       debugDefaultTargetPlatformOverride = TargetPlatform.iOS;
 
       final AdManagerAdRequest request = AdManagerAdRequest(
@@ -1537,6 +1542,7 @@ void main() {
         publisherProvidedId: 'test-pub-id',
         location:
             LocationParams(accuracy: 1.1, longitude: 25, latitude: 38, time: 1),
+        mediationExtrasIdentifier: 'identifier',
         extras: {'key': 'value'},
       );
 
@@ -1550,9 +1556,11 @@ void main() {
       expect(decoded.publisherProvidedId, request.publisherProvidedId);
       expect(decoded.customTargeting, request.customTargeting);
       expect(decoded.customTargetingLists, request.customTargetingLists);
+      // Time is not included on iOS.
       expect(decoded.location!.accuracy, 1.1);
       expect(decoded.location!.longitude, 25);
       expect(decoded.location!.latitude, 38);
+      expect(decoded.mediationExtrasIdentifier, 'identifier');
     });
   });
 }